Add concourse tag to concourse lb vm_extension
- bbl on gcp creates a firewall rule with the target_tag of "concourse",
however this only correctly allows inbound traffic from the internet if
your deployment name is "concourse". bosh will add the deployment name
by default as a tag to each instance, so this worked if your deployment
name is "concourse". moving the tag as a property to the vm_extension
allows an operator to choose an arbitrary name for their concourse
deployment, and traffic will continue to route correctly.